- refactoring the post options into a separate component - fixing tests and build
