+++ frontend/src/components/CreatePost.test.js
@@ -1,12 +1,17 @@
 import { fireEvent, render, screen } from '@testing-library/react';
+import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import CreatePost from './CreatePost';
+import store from '../store';
 
 test('renders MUMBLE link', () => {
   render(
-    <BrowserRouter>
-      <CreatePost />
-    </BrowserRouter>,
+    <Provider store={store}>
+      <BrowserRouter>
+        <CreatePost />
+      </BrowserRouter>
+      ,
+    </Provider>,
   );
   const buttonElement = screen.getByText(/Mumble Now/);
   fireEvent(
